Release sdk-0.4.1
Platform versions: iOS 0.4.1, Android 0.4.1, Flutter 0.2.0, Unity 0.2.0

Automated export from ssp-platform monorepo.

+++ android/DKMadsInterstitialActivity.kt
@@ -194,7 +194,7 @@ class DKMadsInterstitialActivity : Activity() {
                 campaignId = ad.campaignId,
                 creativeId = ad.creativeId ?: ad.id,
                 containerView = videoView,
-                durationMsProvider = { mp.duration.coerceAtLeast(0L) },
+                durationMsProvider = { mp.duration.coerceAtLeast(0).toLong() },
                 currentPositionMsProvider = { videoView.currentPosition.toLong() },
                 isPlayingProvider = { videoView.isPlaying },
                 skippable = true,
